+++ tests/experimental/sync/flexible_sync_tests.cpp
@@ -184,3 +184,55 @@ TEST_CASE("set collection sync", "[set]") {
         CHECK(objs.size() == 1);
     }
 }
+
+TEST_CASE("pause_resume_sync", "[sync]") {
+    auto app = realm::App(realm::App::configuration({Admin::shared().cached_app_id(), Admin::shared().base_url()}));
+
+    SECTION("pause_resume") {
+        auto user = app.login(realm::App::credentials::anonymous()).get();
+
+        auto flx_sync_config = user.flexible_sync_configuration();
+        auto synced_realm = experimental::db(flx_sync_config);
+        auto update_success = synced_realm.subscriptions().update([](realm::mutable_sync_subscription_set &subs) {
+                                                              subs.clear();
+                                                          }).get();
+        CHECK(update_success == true);
+        update_success = synced_realm.subscriptions().update([](realm::mutable_sync_subscription_set &subs) {
+                                                         subs.add<experimental::AllTypesObject>("foo-strings");
+                                                         subs.add<experimental::AllTypesObjectLink>("foo-link");
+                                                     })
+                                 .get();
+        CHECK(update_success == true);
+
+        auto wait_for_state = [](enum realm::sync_session::connection_state expected_state,
+                                 const experimental::db& realm,
+                                 std::function<void(const experimental::db&)> fn) {
+            std::promise<void> p;
+            std::future<void> f = p.get_future();
+            auto token = realm.get_sync_session()->observe_connection_change([&]
+                                                                             (enum realm::sync_session::connection_state,
+                                                                              enum realm::sync_session::connection_state new_state) {
+                if (new_state == expected_state)
+                    p.set_value();
+            });
+            fn(realm);
+
+            if (f.wait_for(std::chrono::milliseconds(10000)) == std::future_status::ready) {
+                realm.get_sync_session()->unregister_connection_change_observer(token);
+                return f.get();
+            } else {
+                throw std::runtime_error("Timeout exceeded");
+            }
+        };
+
+        wait_for_state(realm::sync_session::connection_state::disconnected, synced_realm, [](const experimental::db& realm) {
+            realm.get_sync_session()->pause();
+        });
+        CHECK(synced_realm.get_sync_session()->state() == realm::sync_session::state::paused);
+
+        wait_for_state(realm::sync_session::connection_state::connected, synced_realm, [](const experimental::db& realm) {
+            realm.get_sync_session()->resume();
+        });
+        CHECK(synced_realm.get_sync_session()->state() == realm::sync_session::state::active);
+    }
+}